Moved supermajority functions into new module
